Move algorithms tests to test_logic module
